test(ui): split bootstrap navigation visibility coverage
